try to fix sql 27/01
